Add texthooking variant of xsystem35
The xsystem35-sd2-texthook target creates an AppImage for a version of
xsystem35 with built-in texthooking. This is accomplished by compiling
xsystem35-sdl2 with TEXTHOOK_PRINT defined and piping its output to a
script that reads messages from stdin and copies them to the X
clipboard.

Also added here is the ability to specify which engine to bundle with a
game using the '-e' command line argument, as well as the '-o' argument
which allows specifying the output AppImage file name.